MAINT: refactor compositionspace and constraings into submodule
