+++ smol/moca/composition/__init__.py
@@ -1,0 +1,6 @@
+"""
+Implementation of the CompositionSpace class.
+
+A CompositionSpace represents the space of valid compositions that satisfy given
+constraints, most importantly charge neutrality constraints.
+"""
