+++ tests/test_moca/test_bias.py
@@ -6,7 +6,7 @@ import numpy as np
 import numpy.testing as npt
 import pytest
 
-from smol.moca.composition import get_oxi_state
+from smol.moca.composition.space import get_oxi_state
 from smol.moca.sampler.bias import (
     FugacityBias,
     SquareChargeBias,
